Change phases starts time type to uint256

+++ contracts/AuctionFactory.sol
@@ -8,5 +8,5 @@ contract AuctionFactory {
 
     event AuctionCreated(address auctionContract, address owner, uint numAuctions, address[] allAuctions);
 
-    function createAuction(uint _phaseTwoStart, uint _phaseThreeStart, string memory _description, uint _startingPrice, address payable owner) public { }
+    function createAuction(uint256 _phaseTwoStart, uint256 _phaseThreeStart, string memory _description, uint _startingPrice, address payable owner) public { }
 }
